requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==23.1.0
autobahn==23.6.2
automat==22.10.0
certifi==2025.4.26
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.8
constantly==15.1.0
cryptography==43.0.1
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperlink==21.0.0
idna==3.10
incremental==22.10.0
isodate==0.6.1
logzero==1.7.0
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
pip==25.1
pycparser==2.21
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
pygments==2.19.1
pyotp==2.9.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2023.3
rdflib==6.3.2
rdflib-jsonld==0.6.2
requests==2.31.0
rich==14.0.0
setuptools==78.1.0
shellingham==1.5.4
simplejson==3.19.1
six==1.16.0
smartapi-python==1.4.8
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
twisted==22.10.0
txaio==23.1.1
typer==0.15.3
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.0.3
uvicorn==0.34.2
websocket-client==1.8.0
zope-interface==6.0
upstox-python-sdk==2.17.0